pyproject.toml
[project]
name = "real-estate-investor"
version = "0.1.0"
description = ""
authors = [
    {name = "trojrobert",email = "trojrobert@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.5"
dependencies = [
    "streamlit (>=1.44.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "firecrawl (>=1.15.0,<2.0.0)",
    "folium (>=0.19.5,<0.20.0)",
    "geopy (>=2.4.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
